Significant updates based on several internal build and test cycles. Biggest improvements: Extract by Marker/Sample.
Deploy is getting closer to a stable release by the day. Expect a change to master branch (production release) soon.

-Josh (Developer on the Gobii Project)

+++ gobiiproject/gobii-model/src/main/java/org/gobiiproject/gobiimodel/utils/FileSystemInterface.java
@@ -10,15 +10,23 @@ import static org.gobiiproject.gobiimodel.utils.HelperFunctions.tryExec;
 
 public class FileSystemInterface {
 
+	private static boolean keepAllFiles=false;
 	/**
 	 * As unix RM command.
 	 * @param file
 	 */
 	public static void rm(String file){
+		if(keepAllFiles){
+			return;
+		}
 		HelperFunctions.tryExec("rm "+file);
 	}
 
 	public static void rmIfExist(String file){
+		if(file==null)return;
+		if(keepAllFiles){
+			return;
+		}
 		File f=new File(file);
 		if(f.exists()){
 			f.delete();
@@ -52,5 +60,13 @@ public class FileSystemInterface {
 			ErrorLogger.logError("FileSystemInterface","Unable to call wc",e);
 		}
 		return retVal;
+	}
+
+	/**
+	 * Determines whether or not to keep temporary files. Useful for debuggins
+	 * @param keep
+	 */
+	public static void keepAllFiles(boolean keep){
+		keepAllFiles=keep;
 	}
 }
